tercer avance desplegado a la nube

+++ components/ReservaEdit/ReservaEdit.tsx
@@ -4,6 +4,7 @@ import { Input } from '../ui/input';
 import { Button } from '../ui/button';
 import { Sport } from '../../app/Interface/sport';
 import {API_URL} from "../../config";
+import Swal from "sweetalert2";
 
 interface ReservaMProps {
   field: string;
@@ -53,6 +54,7 @@ export default function ReservaM({
   const [sports, setSports] = useState<Sport[]>([]);
   const [total, setTotal] = useState(0);
   const [price, setPrice] = useState(0);
+  const [phoneError, setPhoneError] = useState('');
 
   //const [booking_id, setbooking_id] = useState(initialData.booking_id);
   const formatTime = (time: string) => {
@@ -61,6 +63,7 @@ export default function ReservaM({
   };
   
   
+   
   const [selectedSportId, setSelectedSportId] = useState<number | null>(null);
   const [customers, setCustomers] = useState<Customer[]>([]);
   const [filteredCustomers, setFilteredCustomers] = useState<Customer[]>([]);
@@ -86,7 +89,8 @@ export default function ReservaM({
       price_morning:number;
       price_evening:number
     }
-    
+ 
+  
     const fetchSports = async () => {
       try {
         const response = await fetch(`${API_URL}/sport`);
@@ -232,14 +236,17 @@ const handleComplete = async () => {
         },
         body: JSON.stringify(requestPayment),
       });
-  
       if (!response.ok) {
-        const errorText = await response.text();
-        console.error('Error response:', errorText);
-        throw new Error('Error al guardar el yape');
+        const errorData = await response.json();
+        if (errorData.message?.includes('user')) {
+          setPhoneError('El número de celular es obligatorio');
+        } else {
+          setPhoneError('');
+        }
+        console.log(`Error al guardar la reserva: `,errorData);
+        return;
       }
   
-      alert('Yape guardado correctamente');
       onClose(); 
     } catch (error) {
       console.error('Error:', error);
@@ -276,35 +283,51 @@ const handleComplete = async () => {
 
   const deleteReserva = async () => {
     try {
-      const response = await fetch(`${API_URL}/booking/${initialData.booking_id}`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      });
+      const result = await Swal.fire({
+        title: "¿Esta seguro que quiere cancelar?",
+        text: "Al cancelar la reserva esta se eliminará permanentemente!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Aceptar",
+        cancelButtonText: "Cancelar",
 
-      if (!response.ok) {
-        const errorText = await response.text();
-        console.error('Error response:', errorText);
-        throw new Error('Error al cancelar la reserva');
+      });
+  
+      if (result.isConfirmed) {
+        const response = await fetch(`${API_URL}/booking/${initialData.booking_id}`, {
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        });
+  
+        if (!response.ok) {
+          const errorText = await response.text();
+          console.error('Error response:', errorText);
+          throw new Error('Error al cancelar la reserva');
+        }
+  
+        const requestPayment2 = {
+          yape: price,
+          user_id,
+          start_time: "",
+          end_time: "",
+          booking_date: "",
+          price: 0,
+          sport_id: 0,
+        };
+  
+        onSave(requestPayment2);
+        onClose(); // Cerrar el modal después de cancelar
       }
-      const requestPayment2={
-        yape:price,
-        user_id,
-      start_time:"",
-      end_time:"",
-      booking_date:"",
-      price:0,
-      sport_id:0,
-      }
-      onSave(requestPayment2)
-      alert('Reserva cancelada correctamente');
-      onClose(); // Cerrar el modal después de cancelar
     } catch (error) {
       console.error('Error:', error);
       alert('Hubo un problema al cancelar la reserva.');
     }
   };
+  
 
   return (
     <Dialog>
@@ -325,7 +348,7 @@ const handleComplete = async () => {
           <p className="mb-2 sm:mb-3 text-[11px] sm:text-sm text-gray-600">cancha: campo {field}</p>
 
           <div className="flex justify-between items-center mb-1 -mt-6 sm:mt-0 ">
-          <div className="relative flex flex-col w-2/3 mt-6">
+          <div className="relative flex flex-col w-2/3 ">
           <label htmlFor="phone" className="text-sm font-medium text-gray-700 mb-2">
             Buscar por teléfono
           </label>
@@ -337,7 +360,9 @@ const handleComplete = async () => {
             className="w-full p-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
             autoComplete="off" // Esto evita la lista por defecto del navegador
           />
-  
+   <p className={`absolute text-red-500 mt-16 text-[10px] ${phoneError ? 'opacity-100' : 'opacity-0'} transition-opacity duration-300 mt-1`}>
+    {phoneError || ' '}
+  </p>
           {phoneSearch && filteredCustomers.length > 0 && phoneSearch.length < 9 && (
             <ul className="absolute left-0 w-full max-h-40 overflow-auto border bg-white mt-16 z-[9999] rounded-md shadow-lg">
               {filteredCustomers.map((customer) => (
@@ -353,17 +378,16 @@ const handleComplete = async () => {
           )}
         </div>
 
-
-          <div className="flex flex-col sm:w-2/3 mt-12 ml-7">
-            <button className="bg-[#3581F2] w-full  h-9 text-white py-1 sm:py-2 p-2 rounded-md hover:bg-blue-800 transition duration-200 flex items-center text-[10px] sm:text-[15px] ">
-              {/* Ícono de agregar */}
+{/* 
+          <div className="flex flex-col sm:w-2/3 mt-7 ml-7">
+            <button  className="bg-[#3581F2] w-full  h-9 text-white py-1 sm:py-2 p-2 rounded-md hover:bg-blue-800 transition duration-200 flex items-center text-[10px] sm:text-[15px] ">
               <svg className="w-3 sm:w-4 h-6 sm:mr-6" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                 <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M12 5v14M5 12h14" />
               </svg>
               Nuevo Cliente
             </button>
           </div>
-
+          */}
             </div>
 
 
@@ -405,6 +429,8 @@ const handleComplete = async () => {
                 onChange={(e) =>{ const newYape = Number(e.target.value);  // Convierte el valor a número
                   setYape(newYape);
                 }}
+                min={initialData.yape}
+                max={price}
                 type="number"
                 className="mt-1 p-2 w-48"
               />
@@ -454,6 +480,10 @@ const handleComplete = async () => {
           </div>
         </div>
       </div>
+
     </Dialog>
+
+
+
   );
 }
